requirements.txt
# Django项目依赖包
# 核心Django框架
Django==4.2.20
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Django扩展
django-jazzmin==3.0.1
django-crispy-forms==2.5
crispy-bootstrap5==2025.6

# 文档处理
python-docx==1.2.0
mammoth==1.11.0
openpyxl==3.1.5
pandas==2.3.3

# Git操作
GitPython==3.1.45

# 火山引擎SDK
volcengine==1.0.206
volcengine-python-sdk==4.0.33

# 环境配置
python-dotenv==1.2.1

# 基础依赖
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2
numpy==2.3.4

# 开发工具
gunicorn==23.0.0

# 测试工具
hypothesis==6.122.3
